requirements.txt
youtube-dl==2021.6.6
discord.py==1.7.3